model_convert/requirements.txt
torch
numpy
hydra-core
omegaconf